+++ src/containers/TabBar/home/UserPostFeed/UserPost.js
@@ -37,7 +37,6 @@ const BottomIconContainer = ({item}) => {
       setLikes(likes + 1);
       getAsyncStorageData("PROFILE_ID").then(profile => {
         addLike(profile, item['id'], item['postType'])
-        console.log(item);
       })
       
     } else {
@@ -52,7 +51,10 @@ const BottomIconContainer = ({item}) => {
 
   const OnPressComments = () => {
     navigation.navigate(StackNav.PostComments,
-      {idPost: item});
+      {
+        idPost: item,
+        fromUser: 'user'
+      });
   }
 
   return (
