+++ src/utils/_support_functions.js
@@ -77,6 +77,7 @@ const transformfPosts = (post) => {
   const new_posts = [];
   const namePost = post['name'];
   const profileImagePost = post['profile_photo'];
+  const descriptionPost = post['description'];
 
   post['post'].forEach(element => {
     const images = [];
@@ -88,6 +89,7 @@ const transformfPosts = (post) => {
 
     const obj = {
       id: element['id'],
+      description: descriptionPost,
       name: namePost,
       profileImage: profileImagePost,
       profileId: profileIdPost,
@@ -100,6 +102,7 @@ const transformfPosts = (post) => {
       views: element['likes'],
       poster: images[0]
     };
+    console.log('OBJ:::', obj)
 
     new_posts.push(obj);
     
@@ -169,6 +172,7 @@ const transformProfileHistory = (post) => {
 
 const transformFeed = (post) => {
   const new_posts = [];
+
   post.forEach(element => {
     const images = [];
     element['multimedia']['data'].forEach(mult => {
@@ -183,7 +187,7 @@ const transformFeed = (post) => {
       profileImage: element['profileImage'],
       text: element['text'],
       image: images,
-      postType: 'POST',
+      postType: element['shareType'],
       // role: 'CEO de empresa', // TODO: delete field
       likes: element['likes'],
       comments: element['comments']
